requirements.txt
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.12.0
grpcio==1.67.1
grpcio-status==1.67.1
ipywidgets==8.1.5
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
langchain==0.3.7
langchain-anthropic==0.2.4
langchain-community==0.3.5
langchain-core==0.3.16
langchain-google-genai==2.0.4
langchain-groq==0.2.1
langchain-mongodb==0.2.0
langchain-text-splitters==0.3.2
langgraph==0.2.45
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.142
numpy==1.26.4
openai==1.54.3
pandas==2.2.3
pymongo==4.10.1
pyparsing==3.2.0
python-dotenv==1.0.1
setuptools==75.1.0
streamlit==1.41.0
streamlit-javascript==0.1.5
streamlit-modal==0.1.2


